Modify DDB for performance bottlenecks module (#30) closes #29
Added DDB provisioned throughput for DDB table performance bottlenecks module. Without this, it defaults to pay per request (on-demand) and scales automatically. Whilst this is a good thing, we want the tracing module to reveal DynamoDB as the bottleneck and we do this by artificially capping the provisioned throughput similar to if a customer set it manually and under-provisioned.